+++ src/test/java/org/springframework/test/service/LoveUService.java
@@ -9,4 +9,6 @@ public interface LoveUService {
   void explode();
 
   String explodeReturn();
+
+  String getName();
 }